requirements.txt
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
coloredlogs==15.0.1
frozenlist==1.3.3
humanfriendly==10.0
idna==3.4
multidict==6.0.4
pyzmq==25.0.0
requests==2.28.2
urllib3==1.26.14
yarl==1.8.2
